Task18 Fixed, Task 20->21

+++ Task18/Domain/Worker.cs
@@ -4,12 +4,12 @@ namespace Task18.Domain;
 
 public class Worker
 {
-    public virtual Int64 Id { get; set; }
+    public virtual int Id { get; set; }
     public virtual string Name { get; set; }
     public virtual string Login { get; set; }
-    public virtual Int32 Age { get; set; }
+    public virtual int Age { get; set; }
     
-    public virtual Int32 Salary { get; set; }
+    public virtual int Salary { get; set; }
     public virtual DateTime Date { get; set; }
 
     public override string ToString()
